Customize "C-a"
lispy.el (lispy-move-beginning-of-line): add and bind to "C-a"

* lispy-test.el (lispy-move-beginning-of-line): Add test.

Previously, when located at a child outline and folding to summary, the
point appeared hidden. It's natural to try to reveal it with "C-a".
Additionally, `lispy-move-beginning-of-line' will jump back to
indentation when already at line start. This can be an additional way of
getting into special.