"e" will lax eval Clojure
* le-clojure.el (lispy--clojure-lax): New defun.
(lispy--eval-clojure): Update.

This is similar to the lax eval for Elisp with "p".

If you mark the expression with a region:

    asdf [1 2 3]

and press "e", you will actually eval this:

    (do (def asdf [1 2 3])
        asdf)

You can do this for let bindings, it's super-useful for debugging.  The
rule is that if the first element of the region is a symbol, and there's
more stuff in the region besides the symbol, a lax eval will be
performed.