lispy.el (lispy--insert-or-call): Don't generate interactive forms.
Since `call-interactively' is used on DEF, the form would be used twice,
which is bad in case e.g. DEF has a `read-char' in its interactive from.
In addition `lispy-eval' can misbehave, since the result
of (help-function-arglist DEF) is in the lexical environment during the
eval - it can shadow the symbol we're trying to eval.