lispy-describe-inline restore window configuration
In case the description is to long to show it in the inline overlay,
lispy-describe-inline calls describe-function/variable which shows the
documentation in an extra help window. However when calling
lispy-decribe-inline again to hide the documentation, the help window is
simply killed, destroying the previous window configuration.

If lispy-restore-windows is set to true, the window configuration will
be saved before the documentation is shown, and restored after it is
hidden again.