Add extra arguments and behavior for lispy-slurp
With an argument of 0, lispy-slurp will slurp as far as possible, and
with an arg of -1, lispy-slurp will slurp until the end or beginning of
the line. In cases where it is not possible to slurp to the end or
beginning of the line, an argument of -1 will slurp as far as possible
on the current line.