Add options to act unsafely in strings or comments
These options only apply when lispy-safe-delete and/or lispy-safe-copy
are non-nil. This is useful since it is not invalid for comments and
strings to contain unmatched delimiters. In some cases comments or
strings may refer to delimiters as characters or contain unmatched
delimiters for some other reason, and this prevents undesirable behavior
in such cases.

Fixes #226