Improve behavior for lispy-pairs when wrapping
Calling lispy-pairs with a negative argument (wrap to the end of the
line) will now simply insert LEFT RIGHT when there is not a wrappable
sexp at the point instead of attempting to wrap the top level or
previous sexp. Also, if no wrapping occurs, a space will no longer
be inserted after the point.