lispy.el (lispy-debug-step-in): Allow mapcar and mapc
* Intended use:

Imagine you have a huge list that's produced elsewhere, and you'd like
to `mapcar' it, but you're yet unsure of this list's element structure.

To rephrase it in code, you start with this:

|(mapcar
  (lambda (x)
    ;; ???
    )
  (subseq
   (ucs-names)
   1024))

Now press "xj" to have `x' become ("THREE RAYS RIGHT" . 128487) and the
point move.

1. You can check the value of `x' with "mie".
2. Or you can move to start implementing the lambda with "d C-m".

From here, it's easy to write down a body such a this, since you have an
example list element that you can play around with:

(mapcar
 (lambda (x)
   (format "%s: %0x" (car x) (cdr x))|)
 (subseq
  (ucs-names)
  1024))

You can readily press "e" and check your work, since `x' already holds
the value that would be the first element of the list that you're
`mapcar'-ing. Assuming the whole list has a homogeneous structure,
checking the result of the whole `mapcar' is optional.