Fix auto wrap commands at list end and in repl
Since looking-back was previously used to check for an empty line,
lispy--not-at-sexp-p would give an incorrect result when used after some
prompt (such as in the minibuffer with eval-expression) causing an
error. Also, since the eol was required in the regexp for the end of
list check, the result would be incorrect at the end of a list followed
by other sexps causing unexpected behavior.

Fixes #325