le-clojure.el (lispy--eval-clojure): Load middleware automatically
On each eval, we check if CIDER is started. If it's not, start it.
Load the middleware in that step as well.

It's possible to edit the middleware in lispy-clojure.clj and reload
it with "C-c C-l" (`cider-load-file'). This will change neither *ns*
nor `lispy--clojure-ns'.

One nitpick with lispy-clojure.clj: when playing around with it, we
remain in the original *ns* (to keep its vars defined). Which means we
often use fully qualified symbols e.g. `lispy-clojure/symbol-function'.

I don't like the alternatives to this: either importing the while
lispy-clojure into the current *ns* or importing the whole
current *ns* into lispy-clojure. Maybe make a new ns and import both
the current *ns* and lispy-clojure there? There's still a potential
for a name clash, but at least it's isolated.