lispy.el (lispy-eval): Refactor
* lispy.el (lispy-eval-alist): Add `clojure-mode'.

* le-clojure.el (lispy-eval-clojure): Add. This is just a sketch,
  which demonstrates how to produce a meaningfull eval from a
  meaningless one using the wrapping context.

Example:

    (->> 5
         |(range)
         (map (fn [x] (* x x)))
         (map (fn [x] (+ x x))))

The highlighted eval of "(range)" is useless by itself. But if we
recognize the context we can turn it into a dwim eval of "(->> 5 (range))".

In the next iteration, the Elisp layer will simply send the whole
parent expression of the current expression as a context for
eval. Then, it's up to the Clojure middleware to make sense of it.

The context will be sent to Clojure as a string for many reasons:

- We don't want to write a Clojure reader in ELisp,

- If we sent an unreadable expression, Clojure can treat it as nil,

- If the Clojure middleware is more isolated from Elisp, it's easier
  to write unit tests