le-clojure.el: "F" now works for defns eval-ed with "e"
Previously, the :file and :line metadata was attached to vars only
when the whole file was loaded by nREPL (e.g. via `cider-load-file').
Without this metadata, the goto-definition
functionality (e.g. `cider-find-var') doesn't work.

Reloading the whole file, just to get goto-definition working for a
newly added defn ins't optimal. It's even worse for notebooks, where
top-level sexps actually do work, instead of just defining things.

What happens now: Elisp sets the file and line position of the sexp to
Clojure via `lispy-clojure/reval'. Which checks if the sexp is a
'(defn ...).  In that case, the `defn` gets additional metadata
`:l-file' and `:l-line' attached to it. We can't use `:file' or
`:line', since those are used by nREPL, and they're set to something
nonsensical like:

    :line 1,
    :file "/tmp/form-init3436469203077247515.clj",

Finally, "F" will call `lispy-clojure/location' that checks if the var
in question has the right metadata. If not, fall back to
`cider-find-var'.