lispy-clojure.clj (object-members): Now a macro
Convert to a macro to make `lispy-clojure-complete-at-point' still
work after the shadows change.

Example:

1. Press "e" on this:

(let [x |(+ 2 2)])

2. Press "C-M-i" on this:

(. x |)

Lispy will eval this Clojure code to get the completions:

    (lispy-clojure/object-members x)

Note that `x' isn't actually a var, it's now a key in `user/shadows'
dictionary. So it's a choice between eval-ing this from Elisp:

    (lispy-clojure/object-members (lispy-clojure/with-shadows x))

Or embedding `with-shadows' into `object-members'.  I prefer the
second approach; it leaks less details about the middleware.

Why it has to be a macro: if it's a function, `x' will be evaluated
before the call to `object-members'. And it's not bound.