Add variable listing modes lispy should not indent in
Do not call indent-for-tab-command from lispy--indent-for-tab if the
current major mode or one of its parent modes is in this list. This
prevents lispy-parens and other commands that use lispy--indent-for-tab
from moving the point to before the beginning of the prompt in
sly-mrepl-mode (#371), geiser-repl-mode (#423), and other major modes
derived from comint-mode.

Fixes #413