Add lispy-show-top-level
Add a new command `lispy-show-top-level' which shows the first line of
the top-level form containing the point in the echo area. This is
useful for seeing which top-level form the point is in when the form
is long enough that its beginning has scrolled off the top of the
screen. The user might not always know which form they're in, e.g.  if
they jumped into the middle of a long form via a search or some other
code navigation technique.

Bind to "xw" inside special mode.

Fixes #555