`lispy-cond<->if-dwim': handle (cond (t XS...))
Previously, a `cond' of the form (cond (t XS...)) would be transformed to (car
XS...), dropping the rest of the else tail.

This is because the `car' of whitespace-trimming `lispy--cases->ifs' was taken,
even though the result was not whitespace + `if' but whitespace + XS....

Fix that by introducing `lispy--progn', like `macroexp-progn' for
`lispy--fast-insert', and to pass the trimmed result trough it instead of `car'.
If the result is a single `if', everything will work, and if the result is a set
of XS... due to an else tail, the result will be a `progn', correctly.