Fix lispy-backward-kill-word delimiter-balancing
* lispy.el (lispy-backward-kill-word): When looking back at ") ", we
would take the first if's then branch because the char before is
whitespace. However, this branch would call backward-kill-word which
doesn't skip over delimiters. Prevent this situation by taking the then
branch only if we're looking back at a word or symbol constituent
followed by whitespace, in which case backward-kill-word behaves
correctly. If we're not looking back at such a pattern, we end up in the
else branch, which does skip over delimiters.

Fixes #584