le-python.el (lispy--eval-python): Add use-in-expr
* Decision
Given this context:
    #+begin_src python
    xs = ["1", "2", "3"]
    x = "2"
    #+end_src

~e~ on src_python{x in xs} should print "True"
~p~ on src_python{x in xs} should give the option to set =x= to become one of the values of =xs=.

* Rationale
src_python{x in xs} is a very common pattern in Python that has two meanings:
1. is the item in the collection
2. iterate over the collection

The split between the two meanings is 50/50, to say roughly. No one meaning prevails over
another.

The default in Python REPL is to choose meaning-1. Previously, lispy would give ~e~
meaning-2. In this way, both bases were covered. If meaning-1 was desired, copying the
code, switching to the REPL and pasting it directly would "work".

This is, however, inconvenient. So we introduce another binding ~p~ to mean "alternative
eval", just like we already have in Elisp and Clojure. And have ~e~ do the usual meaning-1
eval in this case, while ~p~ will do the meaning-2 eval.